refactor(debug): move ClientStats to new Debug section, off legacy /Admin/
Honors memory/architecture/no-admin-url-section.md — /Admin/* is frozen, no
new /Admin/foo routes. Introduces the Debug section as the forward home for
developer/diagnostics pages (logs, db/cache stats, client stats, etc.):

- DebugController at /Debug/* (class-level AdminOnly); ClientStats action and
  its views moved off AdminController / Views/Admin -> Views/Debug
- nav entry repointed Admin -> Debug
- docs/sections/Debug.md section invariants (Status A, greenfield)
- memory/architecture/debug-section.md rule atom + INDEX entry, noting the
  section uses /Debug/* (not /Debug/Admin/*) since it is admin-only end to end
