+++ memory/INDEX.md
@@ -14,6 +14,7 @@ Atomic rules. Fetch the body when the description's trigger matches your task. S
 - [`cached-reads-no-shape-variants`](architecture/cached-reads-no-shape-variants.md) — once a read serves from an in-memory cache of the canonical DTO, do NOT offer `WithEmails` / `IncludeFoo` shape variants. The cache holds one shape — variants reintroduce EF-shaped thinking into a cache-first surface (PR #625 / issue #744).
 - [`consent-record-immutable`](architecture/consent-record-immutable.md) — `consent_records` table: DB triggers block UPDATE/DELETE, INSERT only
 - [`db-enforcement-minimal`](architecture/db-enforcement-minimal.md) — service is the contract, not the DB; only audit-log immutability is doctrinal
+- [`debug-section`](architecture/debug-section.md) — Developer/diagnostics pages live in the **Debug** section at `/Debug/*` (`DebugController`, AdminOnly), not `/Admin/*`. Forward home for logs/db/cache/client stats.
 - [`decorators-talk-only-to-inner`](architecture/decorators-talk-only-to-inner.md) — HARD RULE. A caching/wrapping decorator over interface I may only depend on I (via its keyed inner registration) and the cache plumbing. No sideways repository, service, or sibling-section injections — ever.
 - [`derived-predicates-on-userinfo`](architecture/derived-predicates-on-userinfo.md) — derived user/profile predicates (IsActive, IsStub, NeedsConsentReview, HasRequiredNameFields, etc.) live on `UserInfo`, NOT on `ProfileInfo` or the `Profile` entity. UserInfo is the canonical read surface; ProfileInfo stays a flat field projection; write paths inline.
 - [`display-sort-in-controllers`](architecture/display-sort-in-controllers.md) — display ordering belongs at the presentation layer (controllers / views / view-model assembly), not in services or repositories; repo-layer `OrderBy` allowed only for pagination tie-breakers, top-N, identity-ordered streams (mark with `// arch:db-sort-ok`)
