+++ src/Humans.Web/Controllers/AdminController.cs
@@ -140,51 +140,6 @@ public class AdminController(
         return View(events);
     }
 
-    [HttpGet("ClientStats")]
-    [Authorize(Policy = PolicyNames.AdminOnly)]
-    public IActionResult ClientStats(
-        [FromServices] IClientStatsTracker clientStats,
-        [FromServices] IHttpStatusTracker httpStatus)
-    {
-        var snapshot = clientStats.GetSnapshot();
-
-        static IReadOnlyList<ClientStatRow> ToRows(IReadOnlyList<ClientStatCount> items, long total)
-            => items.Select(i => new ClientStatRow(
-                    i.Label, i.Count, total > 0 ? Math.Round(i.Count * 100.0 / total, 1) : 0))
-                .ToList();
-
-        var statusTotal = httpStatus.Total;
-        var statusRows = httpStatus.GetCounts()
-            .OrderBy(kv => kv.Key)
-            .Select(kv => new HttpStatusRow(
-                kv.Key,
-                StatusCategory(kv.Key),
-                kv.Value,
-                statusTotal > 0 ? Math.Round(kv.Value * 100.0 / statusTotal, 1) : 0))
-            .ToList();
-
-        var vm = new ClientStatsViewModel(
-            TotalPageViews: snapshot.TotalPageViews,
-            OperatingSystems: ToRows(snapshot.OperatingSystems, snapshot.TotalPageViews),
-            Browsers: ToRows(snapshot.Browsers, snapshot.TotalPageViews),
-            DeviceTypes: ToRows(snapshot.DeviceTypes, snapshot.TotalPageViews),
-            TotalResolutionSamples: snapshot.TotalResolutionSamples,
-            Resolutions: ToRows(snapshot.Resolutions, snapshot.TotalResolutionSamples),
-            TotalResponses: statusTotal,
-            StatusCodes: statusRows);
-
-        return View(vm);
-    }
-
-    private static string StatusCategory(int statusCode) => statusCode switch
-    {
-        >= 200 and < 300 => "Success",
-        >= 300 and < 400 => "Redirect",
-        >= 400 and < 500 => "Client error",
-        >= 500 => "Server error",
-        _ => "Other"
-    };
-
     [HttpGet("Maintenance")]
     [Authorize(Policy = PolicyNames.AdminOnly)]
     public IActionResult Maintenance() => View();
