+++ src/Humans.Web/Views/Debug/ClientStats.cshtml
@@ -5,7 +5,7 @@
 
 <div class="d-flex justify-content-between align-items-center mb-3">
     <h1 class="mb-0">Client Statistics</h1>
-    <a asp-action="Index" class="btn btn-outline-secondary btn-sm">
+    <a asp-controller="Admin" asp-action="Index" class="btn btn-outline-secondary btn-sm">
         <i class="fa-solid fa-arrow-left me-1"></i>Admin
     </a>
 </div>
