+++ src/Humans.Web/Views/Debug/_ClientStatTable.cshtml
